requirements.txt
flask
pandas
scikit-learn
numpy
pycryptodome
